fix: disassemble warnings

+++ lib/lpc/program/disassemble.h
@@ -4,5 +4,5 @@
 
 #include "lpc/types.h"
 
-extern void disassemble (FILE *, char *, int, int, program_t *);
+extern void disassemble (FILE *, char *, ptrdiff_t, ptrdiff_t, program_t *);
 extern void dump_line_numbers (FILE *, program_t *);
